Replace deprecated std::char_traits calls with custom char_traits
